Initial sockets based approach
Using lerna + pnpm to setup a monorepo that now has our frontend and websocket backend. All the processing now happens on the backend and it seems to work great for streams themselves (not checked archives) and also will need to actually get it working on the vps
