+++ docker-compose.yml
@@ -1,0 +1,45 @@
+version: '3'
+
+services:
+  frontend:
+    build:
+      context: ./frontend
+      dockerfile: Dockerfile
+    env_file: ./.env
+    environment:
+      - APP_ENV=faker
+      - NODE_OPTIONS=--trace-uncaught
+    ports:
+      - 3000:3000
+    volumes:
+      - haishin-storage:/app/data
+    container_name: haishin-fe
+    restart: unless-stopped
+    networks:
+      - app-network
+
+  backend:
+    build:
+      context: ./backend
+      dockerfile: Dockerfile
+    env_file: ./.env
+    container_name: haishin-be
+    restart: unless-stopped
+    ports:
+      - 8080:8080
+    volumes:
+      - haishin-storage:/app/data
+    networks:
+      - app-network
+
+volumes:
+  haishin-storage:
+    driver: local
+    driver_opts:
+      type: none
+      device: ./data
+      o: bind
+
+networks:
+  app-network:
+    driver: bridge
