+++ package.json
@@ -1,49 +1,12 @@
 {
-  "name": "haishin",
-  "description": "A website that transcribes+translates Japanese streams into English with OpenAI+DeepL",
-  "version": "0.1.0",
+  "name": "root",
   "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "dev:faker": "cross-env APP_ENV=faker next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "postinstall": "patch-package"
-  },
-  "dependencies": {
-    "@dragongoose/streamlink": "^1.1.1",
-    "@ffmpeg/core": "^0.11.0",
-    "@ffmpeg/ffmpeg": "^0.11.6",
-    "@sentry/nextjs": "^7.45.0",
-    "await-exec": "^0.1.2",
-    "date-fns": "^2.29.3",
-    "deepl-node": "^1.8.0",
-    "formidable": "^2.1.1",
-    "next": "13.2.4",
-    "node-ffprobe": "^3.0.0",
-    "openai": "^3.2.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-player": "^2.12.0",
-    "uuid": "^9.0.0"
-  },
   "devDependencies": {
-    "@faker-js/faker": "^7.6.0",
-    "@tailwindcss/aspect-ratio": "^0.4.2",
-    "@types/formidable": "^2.0.5",
-    "@types/node": "18.15.0",
-    "@types/react": "18.0.28",
-    "@types/react-dom": "18.0.11",
-    "@types/uuid": "^9.0.1",
-    "autoprefixer": "^10.4.14",
-    "cross-env": "^7.0.3",
-    "eslint": "8.36.0",
-    "eslint-config-next": "13.2.4",
-    "patch-package": "^6.5.1",
-    "postcss": "^8.4.21",
-    "postinstall-postinstall": "^2.1.0",
-    "tailwindcss": "^3.2.7",
-    "typescript": "4.9.5"
+    "lerna": "^6.6.1"
+  },
+  "pnpm": {
+    "patchedDependencies": {
+      "@ffmpeg/core@0.11.0": "patches/@ffmpeg__core@0.11.0.patch"
+    }
   }
 }
