+++ packages/backend/package.json
@@ -1,0 +1,57 @@
+{
+  "name": "backend",
+  "description": "A website that transcribes+translates Japanese streams into English with OpenAI+DeepL",
+  "version": "2.0.0",
+  "private": true,
+  "main": "dist/app.js",
+  "scripts": {
+    "lint": "eslint . --ext .ts",
+    "build": "tsc -p tsconfig.json",
+    "dev": "cross-env APP_ENV=faker node dist/app.js",
+    "start": "node dist/app.js"
+  },
+  "dependencies": {
+    "@dragongoose/streamlink": "^1.1.1",
+    "@faker-js/faker": "^7.6.0",
+    "@ffmpeg/core": "^0.11.0",
+    "@ffmpeg/ffmpeg": "^0.11.6",
+    "await-exec": "^0.1.2",
+    "date-fns": "^2.29.3",
+    "deepl-node": "^1.9.0",
+    "express": "^4.18.2",
+    "http": "^0.0.1-security",
+    "node-ffprobe": "^3.0.0",
+    "openai": "^3.2.1",
+    "patch-package": "^6.5.1",
+    "socket.io": "^4.6.1",
+    "uuid": "^9.0.0"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.17",
+    "@types/node": "^18.15.11",
+    "@types/uuid": "^9.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.57.0",
+    "@typescript-eslint/parser": "^5.57.0",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.37.0",
+    "typescript": "^5.0.3"
+  },
+  "nx": {
+    "targets": {
+      "dev": {
+        "dependsOn": [
+          "build"
+        ]
+      },
+      "start": {
+        "dependsOn": [
+          "build"
+        ]
+      }
+    }
+  },
+  "optionalDependencies": {
+    "bufferutil": "^4.0.7",
+    "utf-8-validate": "^5.0.10"
+  }
+}
