+++ packages/frontend/.eslintrc.json
