+++ packages/frontend/.gitignore
@@ -1,0 +1,52 @@
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
+# dependencies
+/node_modules
+/.pnp
+.pnp.js
+
+# testing
+/coverage
+
+# next.js
+/.next/
+/out/
+
+# production
+/build
+
+# misc
+.DS_Store
+*.pem
+
+# debug
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+.pnpm-debug.log*
+
+# local env files
+.env*.local
+.env
+
+# vercel
+.vercel
+
+# typescript
+*.tsbuildinfo
+next-env.d.ts
+
+# Streams
+/data/backups/**/*
+/data/streams/**/*
+/data/uploads/**/*
+!/data/backups/.gitkeep
+!/data/streams/.gitkeep
+!/data/uploads/.gitkeep
+
+# Digital Ocean
+/for-server
+
+# Sentry
+.sentryclirc
+next.config.original.js
