+++ packages/frontend/next.config.js
