+++ packages/frontend/pages/api/stream/duration.ts
