+++ packages/frontend/public/favicon.ico
