+++ packages/frontend/sentry.client.config.js
