+++ packages/frontend/sentry.edge.config.js
