+++ packages/frontend/sentry.properties
