+++ packages/frontend/sentry.server.config.js
